payment done

+++ service-service/src/main/java/com/example/service/repository/ServiceRepository.java
@@ -4,5 +4,5 @@ import com.example.service.entity.Service;
 import jdk.jfr.Registered;
 import org.springframework.data.jpa.repository.JpaRepository;
 @Registered
-public interface ServiceRepository extends JpaRepository<Service, Long> {
+public interface ServiceRepository extends JpaRepository<Service, Integer> {
 }
